[Feat] #134 - AppleIdentifier 구현
